规避循环依赖问题
