--- swarmclone/__init__.py
@@ -1,2 +1,0 @@
-from .controller import Controller
-from .modules import *
